refactor compute -> hardware
