+++ CMakeLists.txt
@@ -16,7 +16,7 @@ endif()
 
 # Define the project
 project(
-  xmipp4-compute-metal
+  xmipp4-hardware-metal
 	VERSION ${PROJECT_VERSION} 
 	LANGUAGES C CXX
 )
@@ -27,7 +27,7 @@ include(CTest)
 include(FetchContent)
 
 # Declare build options
-option(XMIPP4_COMPUTE_METAL_LINT_CLANG_TIDY "Lint with clang-tidy" OFF)
+option(XMIPP4_HARDWARE_METAL_LINT_CLANG_TIDY "Lint with clang-tidy" OFF)
 
 # Find xmipp4-core
 if(NOT TARGET xmipp4-core)
@@ -83,7 +83,7 @@ set_target_properties(
 	${TARGET_NAME} 
 	PROPERTIES 
 		CXX_STANDARD 20
-		DEFINE_SYMBOL "XMIPP4_COMPUTE_METAL_EXPORTING"
+		DEFINE_SYMBOL "XMIPP4_HARDWARE_METAL_EXPORTING"
 )
 target_include_directories(
 	${TARGET_NAME}
@@ -107,7 +107,7 @@ target_link_libraries(
 )
 
 # Set up clang-tidy
-if(XMIPP4_COMPUTE_METAL_LINT_CLANG_TIDY)
+if(XMIPP4_HARDWARE_METAL_LINT_CLANG_TIDY)
 	find_program(CLANG_TIDY_EXE clang-tidy REQUIRED)
 	set_target_properties(
 		${TARGET_NAME}
