+++ include/xmipp4/metal/compute/metal_device_backend.hpp
@@ -2,15 +2,15 @@
 
 #pragma once
 
-#include <xmipp4/core/compute/device_backend.hpp>
-#include <xmipp4/core/compute/device_create_parameters.hpp>
+#include <xmipp4/core/hardware/device_backend.hpp>
+#include <xmipp4/core/hardware/device_create_parameters.hpp>
 
 #include <Foundation/Foundation.hpp>
 #include <Metal/Metal.hpp>
 
 namespace xmipp4 
 {
-namespace compute
+namespace hardware
 {
 
 class device_manager;
@@ -37,5 +37,5 @@ public:
 
 }; 
 
-} // namespace compute
+} // namespace hardware
 } // namespace xmipp4
