+++ pyproject.toml
@@ -7,14 +7,14 @@ requires = [
 build-backend = "scikit_build_core.build"
 
 [project]
-name = "xmipp4-compute-metal"
+name = "xmipp4-hardware-metal"
 version = "0.1.0"
 authors = [
     { name = "Oier Lauzirika Zarrabeitia", email = "oierlauzi@bizkaia.eu" },
     { name = "Martín Salinas Antón", email = "ssalinasmartin@gmail.com" },
     { name = "Mikel Iceta Tena", email = "m.icetatena@gmail.com" },
 ]
-description = "METAL compute backend for xmipp4"
+description = "METAL hardware backend for xmipp4"
 classifiers = [
 	"Topic :: Scientific/Engineering :: Image Processing",
     "Intended Audience :: Science/Research",
@@ -33,7 +33,7 @@ dependencies = [
 ]
 
 [project.urls]
-Homepage = "https://github.com/gigabit-clowns/xmipp4-compute-metal"
+Homepage = "https://github.com/gigabit-clowns/xmipp4-hardware-metal"
 
 [tool.scikit-build]
 experimental=true # For wheel.install-dir
