+++ src/metal_plugin_hook.cpp
@@ -4,17 +4,17 @@
 
 #include <xmipp4/core/platform/dynamic_shared_object.h>
 
-#if defined(XMIPP4_COMPUTE_METAL_EXPORTING)
-    #define XMIPP4_COMPUTE_METAL_API XMIPP4_EXPORT
+#if defined(XMIPP4_HARDWARE_METAL_EXPORTING)
+    #define XMIPP4_HARDWARE_METAL_API XMIPP4_EXPORT
 #else
-    #define XMIPP4_COMPUTE_METAL_API XMIPP4_IMPORT
+    #define XMIPP4_HARDWARE_METAL_API XMIPP4_IMPORT
 #endif
 
 static const xmipp4::metal_plugin instance;
 
 extern "C"
 {
-XMIPP4_COMPUTE_METAL_API const xmipp4::plugin* xmipp4_get_plugin() 
+XMIPP4_HARDWARE_METAL_API const xmipp4::plugin* xmipp4_get_plugin() 
 {
     return &instance;
 }
